Provide redirects for old thinking posts
In the past posts within "thinking" were stored as /thinking/post/
whereas now they are under /thinking/year/month/day/post/.  For example,
http://jaisenmathai.com/articles/shuttleworth/ links to
https://shuttleworthfoundation.org/thinking/thinking-fellowship-exit-2014/
which redirects to the homepage now.

Use jekyll-redirect-from to provide redirects for the old posts.  This
is done for posts up to 2016 since the new style was definitely used
as of February 2017 according to archive.org.