Fix aspect ratio of cover images
This will fix the aspect ratio of the cover images for each post. Basically object-fit, will make them a cover and it will auto-adapt the images to fit the height and width of the div ;-) This will work on mobile as well as on desktop and tablets. @jason check it out before merging it.